backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
numpy==2.1.3
Pillow==11.1.0
opencv-python-headless==4.11.0.86
scikit-learn==1.6.1
joblib==1.4.2
psutil==7.0.0
vercel>=0.5.0,<1.0.0
